feat: implement get_channel_name function in lnd_functions; refactor LNDClient error handling and update context manager for invoice logging
