+++ src/context_manager.py
@@ -1,4 +1,5 @@
 import asyncio
+import hashlib
 import json
 from pprint import pprint
 from typing import Any, AsyncGenerator, Dict, Generator, List
@@ -12,6 +13,7 @@ from v4vapp_backend_v2.config import InternalConfig, logger
 from v4vapp_backend_v2.database.db import MyDB
 from v4vapp_backend_v2.lnd_grpc.lnd_client import LNDClient
 from v4vapp_backend_v2.lnd_grpc.lnd_errors import LNDFatalError, LNDSubscriptionError
+from v4vapp_backend_v2.lnd_grpc.lnd_functions import get_channel_name
 from v4vapp_backend_v2.models.htlc_event_models import (
     ChannelName,
     EventType,
@@ -90,31 +92,31 @@ async def subscribe_invoices_loop() -> None:
             raise e
 
 
-async def get_channel_name(channel_id: int) -> ChannelName:
-    if not channel_id:
-        return ChannelName(channel_id=0, name="Unknown")
-    async with LNDClient() as client:
-        request = ln.ChanInfoRequest(chan_id=channel_id)
-        try:
-            response = await client.call(
-                client.lightning_stub.GetChanInfo,
-                request,
-            )
-            chan_info = MessageToDict(response, preserving_proto_field_name=True)
-            pub_key = chan_info.get("node2_pub")
-            if pub_key:
-                response = await client.call(
-                    client.lightning_stub.GetNodeInfo,
-                    ln.NodeInfoRequest(pub_key=pub_key),
-                )
-                node_info = MessageToDict(response, preserving_proto_field_name=True)
-                return ChannelName(
-                    channel_id=channel_id, name=node_info["node"]["alias"]
-                )
-            return ChannelName(channel_id=channel_id, name="Unknown")
-        except Exception as e:
-            logger.exception(e)
-            pass
+# async def get_channel_name(channel_id: int) -> ChannelName:
+#     if not channel_id:
+#         return ChannelName(channel_id=0, name="Unknown")
+#     async with LNDClient() as client:
+#         request = ln.ChanInfoRequest(chan_id=channel_id)
+#         try:
+#             response = await client.call(
+#                 client.lightning_stub.GetChanInfo,
+#                 request,
+#             )
+#             chan_info = MessageToDict(response, preserving_proto_field_name=True)
+#             pub_key = chan_info.get("node2_pub")
+#             if pub_key:
+#                 response = await client.call(
+#                     client.lightning_stub.GetNodeInfo,
+#                     ln.NodeInfoRequest(pub_key=pub_key),
+#                 )
+#                 node_info = MessageToDict(response, preserving_proto_field_name=True)
+#                 return ChannelName(
+#                     channel_id=channel_id, name=node_info["node"]["alias"]
+#                 )
+#             return ChannelName(channel_id=channel_id, name="Unknown")
+#         except Exception as e:
+#             logger.exception(e)
+#             pass
 
 
 async def subscribe_htlc_events() -> AsyncGenerator[HtlcEvent, None]:
@@ -168,7 +170,13 @@ async def subscribe_htlc_events_loop() -> None:
                         "complete": complete,
                     },
                 )
-                # logger.info(tracking.model_dump_json(indent=2))
+                if (
+                    htlc_event.event_type == EventType.RECEIVE
+                    and htlc_event.settle_event
+                    and htlc_event.settle_event.preimage
+                ):
+                    invoice = await lookup_invoice(htlc_event.settle_event.preimage)
+                    print(invoice)
                 if complete:
                     logger.info(f"✅ Complete group, Delete group {htlc_id}")
                     tracking.delete_event(htlc_id)
